% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_process.R
\name{simulate_process}
\alias{simulate_process}
\title{Simulate a path of Cramer–Lundberg model extenssion with capital injections}
\usage{
simulate_process(u = 10, pr = 1, lambda_p = 1, f_p = rexp,
  param_p = list(rate = 1), lambda_n = 1, f_n1 = rexp,
  param_n1 = list(rate = 1/2), f_n2 = actuar::rpareto1,
  param_n2 = list(shape = 3/2, min = 2/3), eps = 0.1,
  max_jumps_number = 10000)
}
\arguments{
\item{u}{a numeric vector of length one specifying an initial capital
(\eqn{u} in the formula representation). Default value is 10.}

\item{pr}{a numeric vector of length one specifying a premium rate (\eqn{c}
in the formula representation). Defaulr value is 1.}

\item{lambda_p}{a numeric vector of length one specifying the rate of a
Poisson process of capital injections. Default value is 1.}

\item{f_p}{a function specifying a random generator for a distribution of
capital injections. Default value is \code{rexp} (an exponential
distribution).}

\item{param_p}{a list of parameters of capital injections' distribution
passed to f_p. The default value is \code{list(rate = 1)}.}

\item{lambda_n}{a numeric vector of length one specifying the rate of a
Poisson process of claims. Default value is 1.}

\item{f_n1}{a function specifying a random generator for a phase-type
distribution of claims. Default value is \code{rexp} (an exponential
distribution).}

\item{param_n1}{a list of parameters of claims' distribution passed to f_n1.
The default value is \code{list(rate = 1 / 2)}.}

\item{f_n2}{a function specifying a random generator for a heavy-tailed
distribution of claims. Default value is \code{actuar::rpareto1} (a Pareto
Type I distribution, from \code{actuar} package).}

\item{eps}{a numeric vector of length one specifying the mixing parameter for
claims' mixed distribution.}

\item{max_jumps_number}{a numeric vector of length one specifying the maximum
number of positive and negative jumps, after which the function stops.
Default value is 10000. Can be set to \code{Inf}.}

\item{param_n1}{a list of parameters of claims' distribution passed to f_n2.
The default value is \code{list(shape = 3 / 2, min = 2 / 3)}.}
}
\value{
An object of class \code{process} which is represented by a list of
elements:
\itemize{
\item path: a matrix of two columns. The first column is the time, the second
is the value of a process.
\item jumps_p a vector of positive (capital injections) jumps' values.
\item time_p a vector indicating times of positive jumps.
\item jumps_n a vector of negative (claims) jumps' values.
\item time_p a vector indicating times of negative jumps.
\item jumps_number a vector indicating the number of all jumps.
}
}
\description{
\code{simulate_process} returns an object of S3 \code{process} class, which
is simulated stochastic process of Cramer–Lundberg model extenssion with
capital injections (i.e. positive jumps).
}
\details{
This function simulates a stochastic process, which follows Cramer–Lundberg
model extenssion with capital injections (i.e. positive jumps) until either
ruin or when maximum number of jumps is attained.
}
\section{Warning}{

The function should be used with a great care when \code{max_jumps_number} is
set to \code{Inf}. In this case a simulating \code{repeat} loop will be
stopped only by ruin. This can cause an infinite loop. The model is as
follows:
\deqn{X_{\epsilon}(t) = u + ct + \sum_{k=1}^{N^{(+)}(t)} U^{(+)}_k -
\sum_{k=1}^{N^{(-)}(t)} U^{(-)}_k}
}

\examples{
\dontrun{
prc <- simulate_process()
}
}
